changes that Ive lost count

+++ pypi_helper.py
@@ -1,9 +1,9 @@
 import hashlib
 import os
 import re
+from pprint import pprint
 
 import requests
-
 from json_helper import *
 
 
@@ -13,20 +13,28 @@ def extract_package_info_dictionary(
     for i_json in search_key_recursive_yield(json_data, "releases"):
         count = 0
         for release_version in iterate_value(i_json):
+            #pprint(release_version)
             flatten_dict = {}
             version_number = list(release_version)[0]
             # Unpack each releases' value
             for specific_release in iterate_value(release_version):
+
                 version = search_key_recursive_return(
                     specific_release, "python_version"
                 )
-                if version == python_version:
-                    break
+                print(version)
+                #if version == python_version:
+                #    break
+
                 type = search_key_recursive_return(specific_release, "packagetype")
-                if type != package_type:
-                    break
+                #if type != package_type:
+                #    break
                 sha256_digest = search_key_recursive_return(specific_release, "sha256")
-                url = search_key_recursive_return(release_version, "url")
+
+                # WHAT THE FUCK! spesific_release not release_version
+                #url = search_key_recursive_return(release_version, "url")
+                url = search_key_recursive_return(specific_release, "url")
+                #pprint(release_version)
                 flatten_dict = {
                     "version_number": version_number,
                     "python_version": version,
@@ -34,6 +42,9 @@ def extract_package_info_dictionary(
                     "sha256_digest": sha256_digest,
                     "url": url,
                 }
+                print(url)
+                #pprint(flatten_dict)
+
                 yield flatten_dict
 
 
@@ -43,15 +54,23 @@ def extract_package_info_dictionary(
 # coverage[toml]
 # requirementslib;
 import parseJson
-def extract_dependency(json_data, extras=True):
+
+
+def extract_dependency(
+    json_data,
+    base_path=None,
+    extras=True,
+):
     dist_set = set()
     requires_dist = search_key_recursive_return(json_data, "requires_dist")
 
-
     if requires_dist is not None:
         for dist in requires_dist:
-            parseJson.save_text(dist, 'requires_dict', '/home/raflman/Documents/gitlab-api')
-            #print(dist)
+            if base_path: parseJson.save_text(dist, "requires_dict", path=base_path)
+
+            p = re.compile("[.0-9a-z-]*", re.IGNORECASE)
+            m = p.match(dist)
+
             splitted_dist = re.split("[\][/!<>=(); ']+", dist)
             if extras is False and "extra" in splitted_dist:
                 break
@@ -63,9 +82,16 @@ def extract_dependency(json_data, extras=True):
             # dist = dist_list[0].split(' ')[0].strip()
             # if the dependency is given as extra. Ex: 'coverage[toml]'
             # for char in filter(dist.endswith, suffix_list): dist = dist.split('[')[0]
-            dist_set.add(splitted_dist[0])
+
+            dist_set.add(m.group())
+            # print('#########################')
+            # print(m.group())
+            # print(splitted_dist[0])
+
+            # dist_set.add(m.group())
     else:
-        print("dist is NONE")
+        pass
+        # print("dist is NONE")
 
     return dist_set
 
@@ -73,6 +99,10 @@ def extract_dependency(json_data, extras=True):
 def download_file(url, sha256_digest=None):
     r = requests.get(url)
     filename = url.rsplit("/", 1)[1]
+    # with open(filename, 'wb') as package:
+    #   package.write(r.content)
+    if os.path.exists(filename):
+        return False
     open(filename, "wb").write(r.content)
     if sha256_digest is not None:
         calculated_digest = calculate_sha256_digest(filename)
@@ -88,8 +118,40 @@ def calculate_sha256_digest(filename):
             sha256_digest.update(byte_block)
     return sha256_digest.hexdigest()
 
+
+def dump_finish_log(iterable, filename, leading_text=None):
+    # timestamp format: xxyyzzjjyy minute, hour, day, month, year
+    file = open(filename, "a")
+    if not type(iterable) is str:
+        for item in iterable:
+            if not item.endswith("\n"):
+                item += "\n"
+            if leading_text:
+                file.write(leading_text)
+            file.write(item)
+    file.close()
+
+
+def log(log_text, log_file, log_dir):
+    log_file = os.path.join(log_dir, log_file)
+    # print("LOGGIN FILE:" + log_file)
+    with open(log_file, "a") as file:
+        if not log_text.endswith("\n"):
+            log_text += "\n"
+        file.write(log_text)
+
+    # if not filename.endswith('.json'):
+    #     filename += '.json'
+    # if not text.endswith('\n'):
+    #     text += '\n'
+    # if path is None:
+    #     path = os.getcwd()
+    # os.chdir(path)
+
+
 def save_json_response(json_data):
-    package_name = 'test'
+    package_name = "test"
+
 
 def if_empty(item):
     return True if item == "" else False
